new protocol header implementation for typed requests in memcache
Summary: This diff includes Header parsing for the new protocol(Caret ^) for typed requests in Umbrella.

Test Plan: tests

Reviewed By: @alikhtarov

Differential Revision: D2306204

+++ mcrouter/lib/network/McServerRequestContext-inl.h
@@ -51,8 +51,9 @@ struct HasDispatchTypedRequest<
 
 template <class OnRequest>
 void McServerOnRequestWrapper<OnRequest>::typedRequestReady(
-  uint64_t typeId, const folly::IOBuf& reqBody,
-  McServerRequestContext&& ctx) {
+    uint32_t typeId,
+    const folly::IOBuf& reqBody,
+    McServerRequestContext&& ctx) {
 
   dispatchTypedRequestIfDefined(
     typeId, reqBody, std::move(ctx),
